+++ apps/api/serializers.py
@@ -3,7 +3,7 @@ from rest_framework_simplejwt.tokens import RefreshToken
 
 from rest_framework import serializers
 
-from apps.core.models import User, Company, Branch, Schedule, WorkDay
+from apps.core.models import User, Company, Branch, Schedule, WorkDay, Telebot
 
 
 class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
@@ -97,3 +97,9 @@ class BranchSerializer(serializers.ModelSerializer):
 
     def get_created_at(self, obj):
         return obj.created_at.__format__('%Y-%m-%d %H:%M:%S')
+
+
+class TelebotSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Telebot
+        fields = '__all__'