+++ apps/api/services/branch_service.py
@@ -1,0 +1,9 @@
+from apps.core.models import Branch
+
+
+def get_branch(user, branch_id):
+    branch = Branch.objects.get(pk=branch_id)
+    if branch.company.owner != user:
+        return False
+    else:
+        return branch