+++ config/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
     path('api/', include('apps.api.urls.swagger_urls', namespace='api_swagger')),
     path('api/company/', include('apps.api.urls.company_urls', namespace='company_endpoints')),
     path('api/branch/', include('apps.api.urls.branch_urls', namespace='branch_endpoints')),
+    path('api/telebot/', include('apps.api.urls.telebot_urls', namespace='telebot_endpoints')),
 ]